refactor: centralize rendering
